--- pepdbagent/pepannot.py
@@ -1,15 +1,0 @@
-from pydantic import BaseModel, Field
-from typing import Optional
-
-
-class Annotation(BaseModel):
-    registry: Optional[str]
-    is_private: Optional[bool]
-    number_of_samples: Optional[int] = Field(alias="n_samples")
-    description: Optional[str]
-    last_update: Optional[str]
-    submission_date: Optional[str]
-
-    class Config:
-        allow_population_by_field_name = True
-        validate_assignment = True
